+++ app/src/main/java/org/androidtown/disfactch/FactCheckingFragment.java
@@ -13,6 +13,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentTransaction;
 
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -22,10 +23,18 @@ import android.widget.RelativeLayout;
 import android.widget.Toast;
 
 import com.dinuscxj.progressbar.CircleProgressBar;
+import com.kakao.sdk.link.LinkClient;
+import com.kakao.sdk.link.model.LinkResult;
+import com.kakao.sdk.template.model.Content;
 import com.kakao.sdk.template.model.FeedTemplate;
+import com.kakao.sdk.template.model.Link;
+import com.kakao.sdk.template.model.TextTemplate;
 
 import java.util.HashMap;
 import java.util.Map;
+
+import kotlin.Unit;
+import kotlin.jvm.functions.Function2;
 
 public class FactCheckingFragment extends Fragment implements CircleProgressBar.ProgressFormatter {
 
@@ -44,6 +53,8 @@ public class FactCheckingFragment extends Fragment implements CircleProgressBar.
     CircleProgressBar cb_fi; // 최종 신뢰도
 
 
+    Button kakao_msg;
+    String url;
 /*
     @Override
     public void onAttach(Context context) {
@@ -106,6 +117,50 @@ public class FactCheckingFragment extends Fragment implements CircleProgressBar.
                 transaction.commit();
             }
         });
+
+        //TODO: 위에서 작성한 link가 string 형식으로 넘어오질 않음
+        url = link.getText().toString();
+        System.out.println(url);
+        System.out.println("===============");
+        kakao_msg = (Button)v.findViewById(R.id.btn_link);
+        kakao_msg.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+
+                /**
+                 * new Link() : 파라미터 순서대로 webLink, mobileLink, androidLink, iosLink
+                 * https://www.beemo.co.kr/entry/카카오-sdk-v2-굳이-자바로-사용하기-2-카카오-링크-메시지?category=743783
+                 */
+                FeedTemplate feedTemplate = new FeedTemplate(new Content("title","imageUrl",    //메시지 제목, 이미지 url
+                        new Link(url),"description",                    //메시지 링크, 메시지 설명
+                        300,300));                                                     //이미지 사이즈
+
+                //TODO: url 넘겨 받는 부분 오류 해결하기
+                TextTemplate textTemplate = new TextTemplate("네이버 링크로 가요 ~" ,new Link(url));
+
+                // line 70dml textTemplate을 feedTemplate으로 바꾸면 feed 형식으로 전송 될 것으로 보임.
+                // 일단 feed로 검사해봐야겠음.
+                LinkClient.getInstance().defaultTemplate(view.getContext(), feedTemplate,null,new Function2<LinkResult, Throwable, Unit>() {
+                    @Override
+                    public Unit invoke(LinkResult linkResult, Throwable throwable) {
+                        if (throwable != null) {
+                            Log.e("TAG", "카카오링크 보내기 실패", throwable);
+                        }
+                        else if (linkResult != null) {
+                            view.getContext().startActivity(linkResult.getIntent());
+
+                            // 카카오링크 보내기에 성공했지만 아래 경고 메시지가 존재할 경우 일부 컨텐츠가 정상 동작하지 않을 수 있습니다.
+                            Log.w("TAG", "Warning Msg: "+ linkResult.getWarningMsg());
+                            Log.w("TAG", "Argument Msg: "+ linkResult.getArgumentMsg());
+                        }
+                        return null;
+                    }
+                });
+            }
+        });
+
+
+
         return v;
     }
