+++ app/src/main/res/layout/fragment_fact_checking.xml
@@ -130,4 +130,12 @@
 
     </RelativeLayout>
 
+    <Button
+        android:id="@+id/btn_link"
+        android:layout_width="wrap_content"
+        android:layout_height="40dp"
+        android:layout_marginLeft="260dp"
+        android:layout_marginBottom="-35dp"
+        android:text="링크" />
+
 </FrameLayout>
